+++ BackendRestAPI/BackendRestAPI/Resources/AssignmentResource.cs
@@ -10,7 +10,7 @@ public sealed class AssignmentResource
 
     public DateOnly CreationDate { get; set; }
 
-    public int ReachablePoints { get; set; }
+    public decimal ReachablePoints { get; set; }
     
     public int SubjectFk { get; set; }
