+++ BackendRestAPI/BackendRestAPI/Resources/SaveAssignmentResource.cs
@@ -8,7 +8,7 @@ public class SaveAssignmentResource
 
     public DateOnly CreationDate { get; set; }
 
-    public int ReachablePoints { get; set; }
+    public decimal ReachablePoints { get; set; }
 
     public int SubjectFk { get; set; }
